improve: Adjust route rules in admin
