--- apps/admin/src/app/(dashboard)/account/loading.tsx
@@ -1,3 +1,0 @@
-import CommonPage from 'src/components/CommonPage';
-
-export default CommonPage.Loading;
