--- apps/admin/src/app/(dashboard)/settings/app-key/loading.tsx
@@ -1,3 +1,0 @@
-import CommonPage from 'src/components/CommonPage';
-
-export default CommonPage.Loading;
