+++ apps/admin/src/guards/AuthGuard/index.ts
@@ -1,3 +1,3 @@
 export { default } from './AuthGuard';
 export { useAuth } from './AuthGuardContext';
-export { auth } from './utils';
+export { auth, getAuthUserId } from './utils';
