+++ apps/service/eslint.config.mjs
@@ -37,6 +37,10 @@ export default tseslint.config(
       '@typescript-eslint/no-unsafe-member-access': 'off',
       '@typescript-eslint/no-unsafe-assignment': 'off',
       '@typescript-eslint/no-unsafe-enum-comparison': 'off',
+      'import/no-extraneous-dependencies': [
+        'error',
+        { packageDir: [__dirname, '../../'] },
+      ],
     },
   },
 );
