+++ apps/service/package.json
@@ -44,7 +44,6 @@
     "handlebars": "^4.7.8",
     "helmet": "^8.0.0",
     "http-proxy-middleware": "^3.0.3",
-    "jsonwebtoken": "^9.0.2",
     "nestjs-zod": "^3.0.0",
     "nodemailer": "^6.10.0",
     "pg": "^8.13.3",
@@ -64,7 +63,6 @@
     "@types/crypto-js": "^4.2.2",
     "@types/express": "^5.0.0",
     "@types/jest": "^29.5.14",
-    "@types/jsonwebtoken": "^9.0.9",
     "@types/multer": "^1.4.12",
     "@types/node": "^22.10.7",
     "@types/nodemailer": "^6.4.17",
