+++ package.json
@@ -15,6 +15,7 @@
   "license": "ISC",
   "devDependencies": {
     "@eslint/eslintrc": "^3",
+    "@types/jsonwebtoken": "^9.0.9",
     "@types/lodash": "^4.17.15",
     "@types/node": "^20",
     "@types/prompts": "^2.4.9",
@@ -31,6 +32,7 @@
     "typescript-eslint": "^8.21.0"
   },
   "dependencies": {
+    "jsonwebtoken": "^9.0.2",
     "lodash": "^4.17.21",
     "zod": "^3.24.2"
   }
